Rely only on trigger level (not frontView position) to determine how to handle a pan gesture
Previously, if the ZUUIRevealController's maxRearViewRevealOverdraw property
was set to 0.0, the user could pan the front view all the way to its rightmost
position, but the front view would immediately animate back to the left.

This change fixes that behavior by relying only on the trigger level, not
the front view's position, to determine how to end the gesture.